feat(timeline): break out into utils
